pyproject.toml
[tool.poetry]
name = "ezotema-bot"
version = "0.1.0"
description = ""
authors = ["Eugene <e.krotov03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = {extras = ["all"], version = "^3.12.0"}
pydantic-settings = "^2.4.0"
typer = "^0.12.4"
jinja2 = "^3.1.4"
sqlalchemy = {extras = ["all"], version = "^2.0.32"}
asyncpg = {extras = ["all"], version = "^0.29.0"}
moviepy = "^1.0.3"
yookassa = "^3.3.0"
email-validator = "^2.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
